rename: setMemberGender 함수를 features 폴더로 이동
