+++ src/entities/member/index.ts
@@ -2,7 +2,6 @@ export {
   getMember,
   setMemberClothesThickness,
   setMemberRegion,
-  setMemberGender,
   setMemberOutingTime,
 } from './api/member';
