+++ src/features/member/api/member.ts
@@ -1,0 +1,12 @@
+import { fetchAPI } from '@/shared/lib';
+
+export async function setMemberGender(gender: string, accessToken: string) {
+  await fetchAPI('/member/gender', {
+    method: 'POST',
+    headers: {
+      Authorization: `Bearer ${accessToken}`,
+      'Content-Type': 'application/json',
+    },
+    body: JSON.stringify({ gender }),
+  });
+}
